# Optimized getMetricsWithLabelNames() helper