**What this PR does**:
While profiling the store-gateway CPU and memory in a cluster running with a large number of tenants per store-gateway instance, I've noticed a small (but significative) % of CPU and allocated objects are due to `getMetricsWithLabelNames()`. In this PR I'm proposing some optimization.

Benchmark:

```
name                         old time/op    new time/op    delta
GetMetricsWithLabelNames-12     931µs ± 1%     225µs ± 2%  -75.86%  (p=0.008 n=5+5)

name                         old alloc/op   new alloc/op   delta
GetMetricsWithLabelNames-12     735kB ± 0%      89kB ± 0%  -87.88%  (p=0.008 n=5+5)

name                         old allocs/op  new allocs/op  delta
GetMetricsWithLabelNames-12     4.03k ± 0%     2.01k ± 0%  -50.10%  (p=0.008 n=5+5)
```

**Which issue(s) this PR fixes**:
N/A

**Checklist**
- [x] Tests updated
- [ ] Documentation added
- [ ] `CHANGELOG.md` updated - the order of entries should be `[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`